requirements.txt
Appium-Python-Client==3.2.0
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
allure-pytest==2.13.2
selenium==4.15.2
